+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.12.5+2
+
+* Declare compatibility with Dart 1.14 and 1.15.
+
+## 0.12.5+1
+
+* Fixed a deadlock bug when using `setUpAll()` and `tearDownAll()`.
+
 ## 0.12.5
 
 * Add `setUpAll()` and `tearDownAll()` methods that run callbacks before and